requirements.txt
Django==4.1.1
djangorestframework==3.14.0
geopy==2.2.0
mysqlclient==2.1.1
python-dotenv==0.21.0
redis==4.3.4

